Super Sanitize
This adds sanitization to the user HTML so that it prevents XSS, bad CSS, and other nastyness. The sanitize gem is amazing by making a great sanitizer!

It also added a custom transform for the text. It basically enforces that all first level tags are p, blockquote, or pre tags, with no floating text. This means that the theme developer can safely assume that the body will have those block tags as the first level tags, regardless of user input.

This also means that user input is pretty flexible. A user can just input raw text and it will paragraph or line break out fine. A user could also enter good HTML paragraphs, and be mostly unscathed (except for sanitization).